// src/enums/wallet.enums.ts
export enum SupportedNetwork {
  BTC = 'Bitcoin',
  BCH = 'BitcoinCash',
  ETH = 'Ethereum',
  LTC = 'Litecoin',
  BSC = 'BinanceSmartChain',
  XRP = 'Ripple',
  XLM = 'Stellar',
  BNB = 'Binance',
  USDT = 'Tether',
  NEO = 'Neo',
  LINK = 'Chainlink',
  TRON = 'Tron',
  DOT = 'Polkadot',
  ADA = 'Cardano',
  DOGE = 'Dogecoin',
  CELO = 'Celo',
  MATIC = 'Polygon',
}

export enum SupportedCurrency {
  BTC = 'BTC',
  BCH = 'BCH',
  LTC = 'LTC',
  CELO = 'CELO',
  KLAY = 'KLAY',
  ONE = 'ONE',
  CUSD = 'CUSD',
  CEUR = 'CEUR',
  ETH = 'ETH',
  ETC = 'ETC',
  XRP = 'XRP',
  XLM = 'XLM',
  DOGE = 'DOGE',
  VET = 'VET',
  NEO = 'NEO',
  BNB = 'BNB',
  BSC = 'BSC',
  CAKE = 'CAKE',
  ZIL = 'ZIL',
  BUSD_BSC = 'BUSD_BSC',
  B2U_BSC = 'B2U_BSC',
  BETH = 'BETH',
  GAMEE = 'GAMEE',
  BBTC = 'BBTC',
  BADA = 'BADA',
  RMD = 'RMD',
  WBNB = 'WBNB',
  BDOT = 'BDOT',
  BXRP = 'BXRP',
  BLTC = 'BLTC',
  BBCH = 'BBCH',
  MATIC = 'MATIC',
  USDC_MATIC = 'USDC_MATIC',
  USDC_BSC = 'USDC_BSC',
  USDC_SOL = 'USDC_SOL',
  USDT = 'USDT',
  GMC = 'GMC',
  GMC_BSC = 'GMC_BSC',
  FLOW = 'FLOW',
  FUSD = 'FUSD',
  USDT_TRON = 'USDT_TRON',
  INRT_TRON = 'INRT_TRON',
  TRON = 'TRON',
  LEO = 'LEO',
  LINK = 'LINK',
  INTENT = 'INTENT',
  EURTENT = 'EURTENT',
  GOLDAX = 'GOLDAX',
  LISK = 'LISK',
  WBTC = 'WBTC',
  UNI = 'UNI',
  SOL = 'SOL',
  FREE = 'FREE',
  MKR = 'MKR',
  USDC = 'USDC',
  MATIC_ETH = 'MATIC_ETH',
  BAT = 'BAT',
  TUSD = 'TUSD',
  BUSD = 'BUSD',
  PAX = 'PAX',
  PLTC = 'PLTC',
  XCON = 'XCON',
  REVV = 'REVV',
  SAND = 'SAND',
  MMY = 'MMY',
  PAXG = 'PAXG',
  HAG = 'HAG',
  ADA = 'ADA',
  XDC = 'XDC',
  LATOKEN = 'LATOKEN',
  USDT_MATIC = 'USDT_MATIC',
  EGLD = 'EGLD',
  ALGO = 'ALGO',
  KCS = 'KCS',
  COIIN = 'COIIN',
  COIIN_BSC = 'COIIN_BSC',
  GLMR = 'GLMR',
  EOS = 'EOS',
  AVAX = 'AVAX',
  FTM = 'FTM',
  ARB = 'ARB',
  OPTIMISM = 'OPTIMISM',
  NEAR = 'NEAR',
  CRO = 'CRO',
  RSK = 'RSK',
  AURORA = 'AURORA',
  GNO = 'GNO',
  DOT = 'DOT',
  KSM = 'KSM',
  OASIS = 'OASIS',
  PALM = 'PALM',
  TEZOS = 'TEZOS',
  ZCASH = 'ZCASH',
  ZEC = 'ZEC',
}

export enum FiatCurrency {
  USD = 'USD',
  EUR = 'EUR',
  GBP = 'GBP',
  TRY = 'TRY',
}